1
Почистил index.js.
Создал отдельный файл с функциями, которые относятся к загрузке популярных фильмов.
Оптимизировал некоторые функции.
Добавил стилизацию кнопкам-переключателям
