+++ src/sass/layout/_movies.scss
@@ -39,17 +39,20 @@
 .popular-btns-list {
   @include d-f;
   justify-content: center;
-  border-bottom: 2px solid #f5f5f5;
+  border-bottom: 2px solid var(--primary-accent-color);
 }
 
 .popular-btn {
   width: 50px;
-  height: 50px;
+  height: 30px;
   border-radius: 10px 10px 0 0;
   outline: none;
   border-color: transparent;
-  background-color: #f5f5f5;
+  background-color: #ececec;
   text-transform: uppercase;
+  transition: background-color var(--hover-focus-transition),
+              color var(--hover-focus-transition),
+              height var(--hover-focus-transition);
   
   &:hover,
   &:focus {
@@ -64,4 +67,5 @@
 .is-active {
   background-color: var(--primary-accent-color);
   color: var(--secondary-text-color);
+  height: 50px;
 }
